+++ examples/run_detection.py
@@ -24,39 +24,41 @@ def get_argument_parser():
 
 def preprocess_fn(x: np.ndarray) -> np.ndarray:
     # TODO: make sure it is correct
-    # x = cv2.resize(src=x, dsize=(416, 416))  # OTX YOLOX
-    x = cv2.resize(src=x, dsize=(992, 736))  # OTX ATSS
+    x = cv2.resize(src=x, dsize=(416, 416))  # OTX YOLOX
+    # x = cv2.resize(src=x, dsize=(992, 736))  # OTX ATSS
     x = x.transpose((2, 0, 1))
     x = np.expand_dims(x, 0)
     return x
 
 
-def main(argv):
+def postprocess_fn(x) -> np.ndarray:
+    """Returns boxes, scores, labels."""
+    return x["boxes"][0][:, :4], x["boxes"][0][:, 4], x["labels"][0]
+
+
+def explain_white_box(args):
     """
     White-box scenario.
-    Insertion of the XAI branch into the Model API wrapper, thus Model API wrapper has additional 'saliency_map' output.
+    Insertion of the XAI branch into the model, thus model has additional 'saliency_map' output.
     """
-
-    parser = get_argument_parser()
-    args = parser.parse_args(argv)
 
     # Create ov.Model
     model: ov.Model
     model = ov.Core().read_model(args.model_path)
 
     # OTX YOLOX
-    # cls_head_output_node_names = [
-    #     "/bbox_head/multi_level_conv_cls.0/Conv/WithoutBiases",
-    #     "/bbox_head/multi_level_conv_cls.1/Conv/WithoutBiases",
-    #     "/bbox_head/multi_level_conv_cls.2/Conv/WithoutBiases",
-    # ]
-    # OTX ATSS
     cls_head_output_node_names = [
-        "/bbox_head/atss_cls_1/Conv/WithoutBiases",
-        "/bbox_head/atss_cls_2/Conv/WithoutBiases",
-        "/bbox_head/atss_cls_3/Conv/WithoutBiases",
-        "/bbox_head/atss_cls_4/Conv/WithoutBiases",
+        "/bbox_head/multi_level_conv_cls.0/Conv/WithoutBiases",
+        "/bbox_head/multi_level_conv_cls.1/Conv/WithoutBiases",
+        "/bbox_head/multi_level_conv_cls.2/Conv/WithoutBiases",
     ]
+    # # OTX ATSS
+    # cls_head_output_node_names = [
+    #     "/bbox_head/atss_cls_1/Conv/WithoutBiases",
+    #     "/bbox_head/atss_cls_2/Conv/WithoutBiases",
+    #     "/bbox_head/atss_cls_3/Conv/WithoutBiases",
+    #     "/bbox_head/atss_cls_4/Conv/WithoutBiases",
+    # ]
 
     # Create explainer object
     explainer = xai.Explainer(
@@ -75,6 +77,7 @@ def main(argv):
     explanation = explainer(
         image,
         targets=[0, 1, 2],  # target classes to explain
+        overlay=True,
     )
 
     logger.info(
@@ -86,6 +89,53 @@ def main(argv):
     if args.output is not None:
         output = Path(args.output) / "detection"
         explanation.save(output, Path(args.image_path).stem)
+
+
+def explain_black_box(args):
+    """
+    Black-box scenario.
+    """
+
+    # Create ov.Model
+    model: ov.Model
+    model = ov.Core().read_model(args.model_path)
+
+    # Create explainer object
+    explainer = xai.Explainer(
+        model=model,
+        task=xai.Task.DETECTION,
+        preprocess_fn=preprocess_fn,
+        postprocess_fn=postprocess_fn,
+        explain_mode=ExplainMode.BLACKBOX,  # defaults to AUTO
+    )
+
+    # Prepare input image and explanation parameters, can be different for each explain call
+    image = cv2.imread(args.image_path)
+
+    # Generate explanation
+    explanation = explainer(
+        image,
+        targets=[0],  # target boxes to explain
+        overlay=True,
+    )
+
+    logger.info(
+        f"Generated {len(explanation.saliency_map)} detection "
+        f"saliency maps of layout {explanation.layout} with shape {explanation.shape}."
+    )
+
+    # Save saliency maps for visual inspection
+    if args.output is not None:
+        output = Path(args.output) / "detection_black_box"
+        explanation.save(output, Path(args.image_path).stem)
+
+
+def main(argv):
+    parser = get_argument_parser()
+    args = parser.parse_args(argv)
+
+    explain_white_box(args)
+    explain_black_box(args)
 
 
 if __name__ == "__main__":